auth config update
